Extract command record routines
