+++ server/src/main/java/org/spine3/server/command/CommandStorage.java
@@ -20,29 +20,21 @@
 
 package org.spine3.server.command;
 
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Function;
-import com.google.common.collect.Iterators;
 import org.spine3.annotations.SPI;
 import org.spine3.base.Command;
-import org.spine3.base.CommandContext;
 import org.spine3.base.CommandId;
 import org.spine3.base.CommandStatus;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
 import org.spine3.server.storage.AbstractStorage;
-import org.spine3.type.TypeName;
 
-import javax.annotation.Nullable;
 import java.util.Iterator;
 
 import static org.spine3.base.CommandStatus.ERROR;
 import static org.spine3.base.CommandStatus.RECEIVED;
-import static org.spine3.base.Commands.generateId;
 import static org.spine3.base.Commands.getId;
-import static org.spine3.base.Identifiers.EMPTY_ID;
-import static org.spine3.base.Identifiers.idToString;
-import static org.spine3.protobuf.Timestamps2.getCurrentTime;
+import static org.spine3.server.command.CommandRecords.newRecordBuilder;
+import static org.spine3.server.command.CommandRecords.toCommandIterator;
 import static org.spine3.validate.Validate.checkNotDefault;
 
 /**
@@ -53,84 +45,8 @@ import static org.spine3.validate.Validate.checkNotDefault;
 @SPI
 public abstract class CommandStorage extends AbstractStorage<CommandId, CommandRecord> {
 
-    private static final Function<CommandRecord, Command> TO_COMMAND =
-            new Function<CommandRecord, Command>() {
-                @Override
-                public Command apply(@Nullable CommandRecord record) {
-                    if (record == null) {
-                        return Command.getDefaultInstance();
-                    }
-                    final Command cmd = record.getCommand();
-                    return cmd;
-                }
-            };
-
     protected CommandStorage(boolean multitenant) {
         super(multitenant);
-    }
-
-    /**
-     * Obtains or generates a {@code CommandId} from the passed command.
-     *
-     * <p>We don't have a command ID in the passed command.
-     * But need an ID to store the error in the record associated
-     * with this command. So, the ID will be generated.
-     *
-     * <p>We pass this ID to the record, so that it has an identity.
-     * But this ID does not belong to the command.
-     *
-     * <p>Therefore, commands without ID can be found by records
-     * where `command.context.command_id` field is empty.
-     */
-    private static CommandId getOrGenerateCommandId(Command command) {
-        CommandId id = getId(command);
-        if (idToString(id).equals(EMPTY_ID)) {
-            id = generateId();
-        }
-        return id;
-    }
-
-    /**
-     * Creates a command storage record builder passed on the passed parameters.
-     *
-     * <p>{@code targetId} and {@code targetIdType} are set to empty strings if
-     * the command is not for an entity.
-     *
-     * @param command            a command to convert to a record. This includes instances of faulty
-     *                           commands. An example of such a fault is missing command ID.
-     * @param status             a command status to set in the record
-     * @param generatedCommandId a command ID to used because the passed command does not have own
-     *                           ID. If the command has own ID this parameter is {@code null}.
-     * @return a storage record
-     */
-    @VisibleForTesting
-    static CommandRecord.Builder newRecordBuilder(Command command,
-                                                  CommandStatus status,
-                                                  @Nullable CommandId generatedCommandId) {
-        final CommandContext context = command.getContext();
-
-        final CommandId commandId = generatedCommandId != null
-                                    ? generatedCommandId
-                                    : context.getCommandId();
-
-        final String commandType = TypeName.ofCommand(command)
-                                           .getSimpleName();
-
-        final CommandRecord.Builder builder =
-                CommandRecord.newBuilder()
-                             .setCommandId(commandId)
-                             .setCommandType(commandType)
-                             .setCommand(command)
-                             .setTimestamp(getCurrentTime())
-                             .setStatus(ProcessingStatus.newBuilder()
-                                                        .setCode(status));
-        return builder;
-    }
-
-    /** Converts {@code CommandStorageRecord}s to {@code Command}s. */
-    @VisibleForTesting
-    static Iterator<Command> toCommandIterator(Iterator<CommandRecord> records) {
-        return Iterators.transform(records, TO_COMMAND);
     }
 
     /**
@@ -156,7 +72,9 @@ public abstract class CommandStorage extends AbstractStorage<CommandId, CommandR
     protected void store(Command command, CommandStatus status) {
         checkNotClosed();
 
-        final CommandRecord record = newRecordBuilder(command, status, null).build();
+        final CommandRecord record = newRecordBuilder(command,
+                                                      status,
+                                                      null).build();
         final CommandId commandId = getId(command);
         write(commandId, record);
     }
@@ -173,7 +91,7 @@ public abstract class CommandStorage extends AbstractStorage<CommandId, CommandR
         checkNotClosed();
         checkNotDefault(error);
 
-        final CommandId id = getOrGenerateCommandId(command);
+        final CommandId id = CommandRecords.getOrGenerateCommandId(command);
 
         final CommandRecord.Builder builder = newRecordBuilder(command, ERROR, id);
         builder.getStatusBuilder()
